Move toolbar to own file, tighter padding when collapsed
